pyproject.toml
[project]
name = "translation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.0.0",
    "aiohttp==3.8.6",
    "aiosignal==1.3.1",
    "appdirs==1.4.4",
    "asttokens>=3.0.0",
    "async-timeout==4.0.3",
    "attrs==23.1.0",
    "backcall>=0.2.0",
    "backports-functools-lru-cache>=2.0.0",
    "cachetools==5.3.1",
    "certifi==2023.7.22",
    "charset-normalizer==3.2.0",
    "click==8.1.7",
    "cmake==3.27.2",
    "comm>=0.2.3",
    "contourpy==1.1.0",
    "cycler==0.11.0",
    "datasets==2.14.5",
    "debugpy>=1.8.17",
    "decorator>=5.2.1",
    "dill==0.3.7",
    "docker-pycreds==0.4.0",
    "entrypoints>=0.4",
    "evaluate==0.4.1",
    "executing>=2.2.1",
    "filelock==3.12.3",
    "fonttools==4.42.1",
    "frozenlist==1.4.0",
    "fsspec==2023.6.0",
    "gitdb==4.0.10",
    "gitpython==3.1.37",
    "google-auth==2.23.1",
    "google-auth-oauthlib==1.0.0",
    "grpcio==1.58.0",
    "huggingface==0.0.1",
    "huggingface-hub==0.18.0",
    "idna==3.4",
    "importlib-metadata==6.8.0",
    "importlib-resources==6.0.1",
    "ipykernel>=7.0.1",
    "ipython>=9.6.0",
    "jedi>=0.19.2",
    "jinja2==3.1.2",
    "joblib==1.3.2",
    "jupyter-client>=8.6.3",
    "jupyter-core>=5.9.1",
    "kiwisolver==1.4.5",
    "lit==16.0.6",
    "markdown==3.4.4",
    "markupsafe==2.1.3",
    "matplotlib==3.7.3",
    "matplotlib-inline>=0.2.1",
    "mpmath==1.3.0",
    "multidict==6.0.4",
    "multiprocess==0.70.15",
    "nest-asyncio>=1.6.0",
    "networkx==3.1",
    "numpy==1.24.4",
    "oauthlib==3.2.2",
    "onnx==1.14.1",
    "packaging>=25.0",
    "pandas==2.0.3",
    "parso>=0.8.5",
    "pathtools==0.1.2",
    "pexpect>=4.9.0",
    "pickleshare>=0.7.5",
    "pillow==10.0.0",
    "prompt-toolkit>=3.0.52",
    "protobuf==4.24.3",
    "psutil>=7.1.1",
    "ptyprocess>=0.7.0",
    "pure-eval>=0.2.3",
    "pyarrow==13.0.0",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "pygments>=2.19.2",
    "pyparsing==3.1.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv==1.0.0",
    "pytz==2023.3",
    "pyyaml==6.0.1",
    "pyzmq>=27.1.0",
    "regex==2023.8.8",
    "requests==2.31.0",
    "requests-oauthlib==1.3.1",
    "responses==0.18.0",
    "rsa==4.9",
    "ruff>=0.14.2",
    "sacremoses==0.0.53",
    "sentencepiece==0.1.99",
    "sentry-sdk==1.31.0",
    "setproctitle==1.3.2",
    "six>=1.17.0",
    "smmap==5.0.1",
    "stack-data>=0.6.3",
    "sympy==1.12",
    "tensorboard==2.14.0",
    "tensorboard-data-server==0.7.1",
    "torch==2.1.0",
    "torchinfo==1.8.0",
    "torchtext>=0.16.0",
    "torchvision==0.16.0",
    "tornado>=6.5.2",
    "tqdm==4.66.1",
    "traitlets>=5.14.3",
    "typing-extensions>=4.15.0",
    "tzdata==2023.3",
    "urllib3==2.0.5",
    "wcwidth>=0.2.14",
    "werkzeug==2.3.7",
    "xxhash==3.4.1",
    "yarl==1.9.2",
    "zipp==3.16.2",
]
